Add image to db without relations for now.

+++ src/models/_entities/images.rs
@@ -18,7 +18,7 @@ pub struct Model {
     #[sea_orm(column_type = "Float", nullable)]
     pub duration: Option<f32>,
     pub format: String,
-    pub size_bytes: i32,
+    pub size_bytes: i64,
     pub datetime_local: DateTime,
     pub datetime_utc: Option<DateTime>,
     pub datetime_source: String,
